Standardise labels
Source: R/clean_labels.R

This function standardises labels e.g. used as variable names or character string values, removing non-ascii characters, replacing diacritics (e.g. é, ô) with their closest ascii equivalents, and standardises separating characters. See details for more information on label transformation.

clean_labels(x, sep = "_")

Arguments

x
A vector of labels, normally provided as characters.

sep
A character string used as separator, defaulting to '_'.

Details

The following changes are performed:

all non-ascii characters are removed

all diacritics are replaced with their non-accentuated equivalents, e.g. 'é', 'ê' and 'è' become 'e', 'ö' becomes 'o', etc.

all characters are set to lower case

separators are standardised to the use of a single character provided in sep (defaults to '_'); heading and trailing separators are removed.

Examples

clean_labels("-_-This is; A    WeÏrD**./sêntënce...")

#> [1] "this_is_a_weird_sentence"

clean_labels("-_-This is; A    WeÏrD**./sêntënce...", sep = ".")

#> [1] "this.is.a.weird.sentence"

input <- c("Peter and stëven", "peter-and.stëven", "pëtêr and stëven  _-")
input

#> [1] "Peter and stëven"     "peter-and.stëven"     "pëtêr and stëven  _-"

clean_labels(input)

#> [1] "peter_and_steven" "peter_and_steven" "peter_and_steven"