Anonymise data using SHA1

This function uses SHA1 algorithm to anonymise data, based on user-indicated data fields. Data fields are concatenated first, then each entry is hashed. The function can either return a full detailed output, or short labels ready to use for 'anonymised data'.

hash_names(..., size = 6, full = TRUE)

Arguments

...
Data fields to be hashed.

size
The number of characters retained in the hash.

full
A logical indicating if the a full output should be returned as a data.frame, including original labels, shortened hash, and full hash.

Details

Once concatenated (using "_" as a separator), the labels are modified as follows:

all spaces and non-alphanumeric characters are removed

all diacritics are replaced with their non-accentuated equivalents, e.g. 'é', 'ê' and 'è' become 'e', 'ö' becomes 'o', etc.

all non-ascii characters are removed

all characters are set to lower case

Note that the algorithm is not 'salted'. This can be done by tweaking the input characters before the hashing.

Examples

first_name <- c("Jane", "Joe", "Raoul")
last_name <- c("Doe", "Smith", "Dupont")
age <- c(25, 69, 36)

hash_names(first_name, last_name, age)

#>           label hash_short                                     hash
#> 1     janedoe25     274be0 274be0e34366ab2b8798adb31ad161ea8441c410
#> 2    joesmith69     4c20ad 4c20ad4911e4941eaa3c12023440da406eaf3353
#> 3 raouldupont36     4573ee 4573ee25d7f395d94a89c2fea71bfed27933b6d9

hash_names(first_name, last_name, age,
           size = 8, full = FALSE)

#> [1] "274be0e3" "4c20ad49" "4573ee25"