Simulate simple linelist data

This function simulates a simple linelist data including dates of epidemiological events and basic patient information. No underlying epidemiological model is used.

sim_linelist(
  n = 1,
  onset_from = as.Date("2020-01-01"),
  onset_span = 60,
  report_delay = 7,
  cfr = 0.1
)

Arguments

n

Number of entries to simulate.

onset_from

The earliest date of symptom onset which can be generated.

onset_span

The time span over which to generate dates of onset.

report_delay

The mean delay between onset and reporting, using a Poisson distribution.

cfr

The case fatality ratio, i.e. the proportion of patient dying from the infection (used to generate the 'outcome' variable).

Author

Thibaut Jombart thibautjombart@gmail.com

Examples

sim_linelist(10)
#>    id date_of_onset date_of_report gender  outcome
#> 1   1    2020-02-08     2020-02-16 female recovery
#> 2   2    2020-01-31     2020-02-11   male recovery
#> 3   3    2020-01-30     2020-02-06   male recovery
#> 4   4    2020-02-22     2020-02-29   male recovery
#> 5   5    2020-01-24     2020-02-02   male recovery
#> 6   6    2020-01-05     2020-01-09   male recovery
#> 7   7    2020-01-09     2020-01-20   male recovery
#> 8   8    2020-01-18     2020-01-22 female recovery
#> 9   9    2020-02-19     2020-02-27   male recovery
#> 10 10    2020-01-12     2020-01-18 female recovery