fit_disc_gamma {epitrix}	R Documentation

Fit discretised distributions using ML

Description

These functions performs maximum-likelihood (ML) fitting of a discretised distribution. This is typically useful for describing delays between epidemiological events, such as incubation period (infection to onset) or serial intervals (primary to secondary onsets). The function optim is used internally for fitting.

Usage

fit_disc_gamma(x, mu_ini = 1, cv_ini = 1, interval = 1, w = 0, ...)

Arguments

x
A vector of numeric data to fit; NAs will be removed with a warning.

mu_ini
The initial value for the mean 'mu', defaulting to 1.

cv_ini
The initial value for the coefficient of variation 'cv', defaulting to 1.

interval
The interval used for discretisation; see distcrete.

w
The centering of the interval used for discretisation; see distcrete.

...
Further arguments passed to optim.

Value

The function returns a list with human-readable parametrisation of the discretised Gamma distibution (mean, sd, cv), convergence indicators, and the discretised Gamma distribution itself as a distcrete object (from the distcrete package).

Author(s)

Thibaut Jombart thibautjombart@gmail.com

See Also

The distcrete package for discretising distributions, and optim for details on available optimisation procedures.

Examples

## generate data

mu <- 15.3 # days
sigma <- 9.3 # days
cv <- mu / sigma
cv
param <- gamma_mucv2shapescale(mu, cv)

if (require(distcrete)) {
w <- distcrete("gamma", interval = 1,
               shape = param$shape,
               scale = param$scale, w = 0)

x <- w$r(100)
x

fit_disc_gamma(x)
}

[Package epitrix version 0.2.2 Index]