Anonymise data using SHA1
Source: R/hash_names.R

This function uses SHA1 algorithm to anonymise data, based on user-indicated data fields. Data fields are concatenated first, then each entry is hashed. The function can either return a full detailed output, or short labels ready to use for 'anonymised data'. Before concatenation (using "_" as a separator) to form labels, inputs are modified using clean_labels.

hash_names(..., size = 6, full = TRUE, salt = NULL)

Arguments

...
Data fields to be hashed.

size
The number of characters retained in the hash.

full
A logical indicating if the a full output should be returned as a data.frame, including original labels, shortened hash, and full hash.

salt
An optional object to be used to 'salt' the hashing algorithm (see details). Ignored if NULL (default).

Details

The argument salt can be used for salting the algorithm, i.e. adding an extra input to the input fields (the 'salt') to change the resulting hash and prevent identification of individuals via pre-computed hash tables. Objects provided as salt will themselves be hashed using SHA1 algorithm, and the full hash is appended to the labels.

See also

Examples

first_name <- c("Jane", "Joe", "Raoul")
last_name <- c("Doe", "Smith", "Dupont")
age <- c(25, 69, 36)

hash_names(first_name, last_name, age)

#>             label hash_short                                     hash
#> 1     jane_doe_25     4ea703 4ea703cdc94049fc649cd152faa5beed02858d69
#> 2    joe_smith_69     c76ee2 c76ee26d66438710023bb256a644ba9db9c2f6be
#> 3 raoul_dupont_36     abdf0b abdf0bf7990ce5127b406ca16fc1edd659ff51b3

hash_names(first_name, last_name, age,
           size = 8, full = FALSE)

#> [1] "4ea703cd" "c76ee26d" "abdf0bf7"

## salting the hashing (more secure!)
hash_names(first_name, last_name) # unsalted - less secure

#>          label hash_short                                     hash
#> 1     jane_doe     979145 979145053f9208e587138eede409df465ba59dfd
#> 2    joe_smith     d7cd9c d7cd9ca61ac89d6a91bba7d10d76138627580c72
#> 3 raoul_dupont     7aaaa6 7aaaa6fb319d92d8875f1d0805c9e5365fef1699

hash_names(first_name, last_name, salt = 123) # salted with an integer

#>          label hash_short                                     hash
#> 1     jane_doe     d7d9fd d7d9fdc9c0c40595ccb8c2bcadb3d31442b68f88
#> 2    joe_smith     86f115 86f115c5011d7ffd5bf701e7490fc37f0295e252
#> 3 raoul_dupont     97ceca 97cecafeb33f1c1f14ababf454fce2283f9395c2

hash_names(first_name, last_name, salt = "foobar") # salted with an character

#>          label hash_short                                     hash
#> 1     jane_doe     1e4d58 1e4d582f239490fd161d35b15ff98394654eb501
#> 2    joe_smith     01688c 01688c1d85afd0b433ebcfbc26286b730b56ea1d
#> 3 raoul_dupont     e165c5 e165c59225b70cdef28564260d9175336050a40c